[bug]: fix potential memory overflow
bit_nset is used to set bits [start, stop], which the stop position
is included, so we should pass `bitstring size - 1` as the stop,
or it will potentially overflow. It is ok with the month and dom,
because their allocted byte was not fully used, but for dow and
hour, they use 8 bits and 24bits, the following statement will
write the next byte.

bit_nset(e->hour, 0, (LAST_HOUR-FIRST_HOUR+1));
bit_nset(e->dow, 0, (LAST_DOW-FIRST_DOW+1));

Signed-off-by: Junwang Zhao <zhjwpku@gmail.com>